feat(WPLoader): add debug information to WPLoader sudden dies to guide debug
When used in conjuction with the WPDb/Db module the WPLoader module might find, especially on first
runs, an empty database and silently die/exit. In these cases it's very difficult for users to find
out what happened and the additional messaging tries to provide help. Ideally the WPLoader module
should detect such a situation but, due to low reports on the issue, it does not seem worth to put
in place a more complex solution that could hamper users' abilities to use the modules.