fix(WPLoader) avoid eager loading of factories
In `2.2.16` I've exposed the `FactoryStore` on the `$tester` property in
integration tests too. That change caused an eager loading of the
factory classes by the `FactoryStore` class that would cause Warnings
due to some accessory values (globals) not being loaded yet. This commit
makes the instantiation of each factory, a property of the
`FactoryStore` object, lazy load.

fix #281