refactor(sep. proc. support) remove const. definition
Remove the definition of the `PHPUNIT_COMPOSER_INSTALL` constant to rely
on the definition, and following use by PHPUnit, of the
`__PHPUNIT_BOOTSTRAP` global.

This addresses issues like #563 where the definition of the PHPUnit
constant before time would cause double definition issues difficult to
debug, especially in the context of applications using muted PHP
processes.