feat: allow preventing inheriting select env vars.
This is very useful to not the arguments passed
to codeception influence the arguments passed to the wp-cli command.

An example, running: codecept run unit --quiet
Will, among other populate a global SHELL_VERBOSITY variable. This is probably not something we
want to inherit, as now our runner influences
out test state.