feat(PHPUnit) bare minimum PHPUnit 10/11 support
The Core PHPUnit suite supports, at most, PHPUnit 9.5.
This code scaffolds a first attempt at supporting PHPUnit version 10 and
11, the curent latest one, adding as little destruction as possible.

This will not likely deal with all the intricacies of it, but will open
the door to supporting the latest versions of Codeception and its
modules supporting the latest versions of PHPUnit.