fix(Module/WPLoader) load WordPress in _beforeSuite method
fixes #744

Change the code of the `WPLoader` module to load WordPress, whether the
`loadOnly` flag is set to `true` or `false`, in the `_beforeSuite`
method.

This removes the need, for the module, to rely on the dispatching,
subscribing and need to do so, to the main Codeception system through
the Codeception event dispatcher.

Kudos to @lxbdr for the proposed solution.