refactor(LoadSandbox) move debug information to exception
Since the output of `codecept_debug` will only show if the user is
running the Codeception command using the `--debug` output, I've moved
the information that will help the user debug the issue to the exception
to make sure it will always print. I've changed the wording a bit to be
more general and cover other commands (e.g. the `console` one) that will
use the LoadSandbox class and might fail like the `run` command does.
Fix test.