Use fallback for `$_composer_autoload_path`
There are cases where this global may be undefined, and Composer recommends
having a fallback. Without one, method signature is incorrect and usages like the
`vendorDir()` method will fail if they expect a string.

The fallback first looks for `vendor` in the same directory as `codeception.yml`,
and if that's not found it guesses what the autoload path _ought_ to be.

More info: https://getcomposer.org/doc/articles/vendor-binaries.md#finding-the-composer-autoloader-from-a-binary